ingestion/requirements.txt
requests
pandas
psycopg2-binary
redis
python-dotenv
schedule
pyproj
shapely
